Commit con avances en la implemenación de las victorias y salon de la fama
