--- src/edu/katheryn_azano/evidencia3/data/Simbolos.java
@@ -1,4 +1,0 @@
-package edu.katheryn_azano.evidencia3.data;
-
-public class Simbolos {
-}
